Program to calculate the thermal transmittance of a building fabric
This is a simple program designed to calculate the thermal transmittance of a building fabric (also called a "U-value", be it a wall, floor or a roof.
It can evaluate fabrics having no thermal bridging, or layers with thermal bridging, for example, you might often come across timber battens spaces at regular intervals, with insulation between them.
The program can also write an HTML file which shows the workings of how the final U-value was reached.